+++ src/io/write_hdf5.cpp
@@ -44,11 +44,11 @@ namespace io {
         H5Gclose(H5Gcreate(root_file_id, "PartType6", H5P_DEFAULT, H5P_DEFAULT, H5P_DEFAULT));
 
         size_t particle_data_width = 0;
-        for (int i = 0; i < sim_data.labels.size(); i++) { 
+        for (size_t i = 0; i < sim_data.labels.size(); i++) { 
             particle_data_width += sim_data.dim[i];
         }
 
-        for (int i = 0; i < sim_data.labels.size(); i++) {
+        for (size_t i = 0; i < sim_data.labels.size(); i++) {
             auto& label = sim_data.labels[i];
             auto& dim = sim_data.dim[i];
 
@@ -59,8 +59,8 @@ namespace io {
             
             std::vector<double> buffer(particle_count * dim);
 
-            for (int ii = 0; ii < particle_count; ii++) {
-                for (int iii = 0; iii < dim; iii++) {
+            for (size_t ii = 0; ii < particle_count; ii++) {
+                for (unsigned short iii = 0; iii < dim; iii++) {
                     buffer[ii + iii] = sim_data.data[ii * particle_data_width + i + iii]; // iith particle, ith force + iiith component 
                 }
             }
