+++ src/io/write_hdf5_dump.cpp
@@ -19,7 +19,7 @@ namespace io {
         const size_t particle_count = sim_data.particle_count;
 
         size_t particle_data_width = 0;
-        for (int i = 0; i < sim_data.labels.size(); i++) { 
+        for (size_t i = 0; i < sim_data.labels.size(); i++) { 
             particle_data_width += sim_data.dim[i];
         }
 
@@ -95,14 +95,14 @@ namespace io {
         double scalar_buffer[particle_count];
         size_t idx = 0;
 
-        for (int p = 0; p < particle_count; p++) {
+        for (size_t p = 0; p < particle_count; p++) {
             scalar_buffer[idx++] = sim_data.data[(p * particle_data_width) + pos_idx];
         }
         write_dataset(group_id, "x", 1, new hsize_t[1]{particle_count}, H5T_IEEE_F64BE, H5T_NATIVE_DOUBLE,
                       scalar_buffer);
 
         idx = 0;
-        for (int p = 0; p < particle_count; p++) {
+        for (size_t p = 0; p < particle_count; p++) {
             scalar_buffer[idx++] = sim_data.data[(p * particle_data_width) + pos_idx + 1];
         }
         write_dataset(group_id, "y", 1, new hsize_t[1]{particle_count}, H5T_IEEE_F64BE, H5T_NATIVE_DOUBLE,
@@ -110,14 +110,14 @@ namespace io {
 
 #if DIM == 3
         idx = 0;
-        for (int p = 0; p < particle_count; p++) {
+        for (size_t p = 0; p < particle_count; p++) {
             scalar_buffer[idx++] = sim_data.data[(p * particle_data_width) + pos_idx + 2];
         }
         write_dataset(group_id, "z", 1, new hsize_t[1]{particle_count}, H5T_IEEE_F64BE, H5T_NATIVE_DOUBLE,
                       scalar_buffer);
 
         idx = 0;
-        for (int p = 0; p < particle_count; p++) {
+        for (size_t p = 0; p < particle_count; p++) {
             scalar_buffer[idx++] = sim_data.data[(p * particle_data_width) + vel_idx + 2];
         }
         write_dataset(group_id, "vz", 1, new hsize_t[1]{particle_count}, H5T_IEEE_F64BE, H5T_NATIVE_DOUBLE,
@@ -125,28 +125,28 @@ namespace io {
 #endif
 
         idx = 0;
-        for (int p = 0; p < particle_count; p++) {
+        for (size_t p = 0; p < particle_count; p++) {
             scalar_buffer[idx++] = sim_data.data[(p * particle_data_width) + vel_idx];
         }
         write_dataset(group_id, "vx", 1, new hsize_t[1]{particle_count}, H5T_IEEE_F64BE, H5T_NATIVE_DOUBLE,
                       scalar_buffer);
 
         idx = 0;
-        for (int p = 0; p < particle_count; p++) {
+        for (size_t p = 0; p < particle_count; p++) {
             scalar_buffer[idx++] = sim_data.data[(p * particle_data_width) + vel_idx + 1];
         }
         write_dataset(group_id, "vy", 1, new hsize_t[1]{particle_count}, H5T_IEEE_F64BE, H5T_NATIVE_DOUBLE,
                       scalar_buffer);
 
         idx = 0;
-        for (int p = 0; p < particle_count; p++) {
+        for (size_t p = 0; p < particle_count; p++) {
             scalar_buffer[idx++] = sim_data.data[(p * particle_data_width) + rho_idx];
         }
         write_dataset(group_id, "rho", 1, new hsize_t[1]{particle_count}, H5T_IEEE_F64BE, H5T_NATIVE_DOUBLE,
                       scalar_buffer);
 
         idx = 0;
-        for (int p = 0; p < particle_count; p++) {
+        for (size_t p = 0; p < particle_count; p++) {
             scalar_buffer[idx++] = sim_data.data[(p * particle_data_width) + prr_idx];
         }
         write_dataset(group_id, "p", 1, new hsize_t[1]{particle_count}, H5T_IEEE_F64BE, H5T_NATIVE_DOUBLE,
